fix uniform

+++ src/my_engine/my_utils/my_color.h
@@ -3,9 +3,9 @@
 #include "my_utils/log.h"
 
 inline std::array<float, 4> GetColorVec4(uint32_t argb) {
-    float r = (argb >> 16) & 0xff;
-    float g = (argb >> 8) & 0xff;
-    float b = (argb >> 0) & 0xff;
-    float a = (argb >> 24) & 0xff;
+    float r = (float) ((argb >> 16) & 0xff);
+    float g = (float) ((argb >> 8) & 0xff);
+    float b = (float) ((argb >> 0) & 0xff);
+    float a = (float) ((argb >> 24) & 0xff);
     return { r / 255, g / 255, b / 255, a / 255, };
 }
